# Optimize UAST transforms