Address performance issues in the last SDK version.

In short: large transforms are an order of magnitude faster now. 

```
benchmark                         old ns/op      new ns/op     delta
BenchmarkGoDriver/transform-4     3181055629     261985232     -91.76%

benchmark                         old allocs     new allocs     delta
BenchmarkGoDriver/transform-4     18913347       561193         -97.03%

benchmark                         old bytes      new bytes     delta
BenchmarkGoDriver/transform-4     2509354920     54671790      -97.82%
```